+++ experimental/CHANGELOG.md
@@ -10,6 +10,8 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 
 ### :rocket: Features
 
+* feat(otlp-grpc-exporter): add gRPC channelOptions as config parameter [#6332](https://github.com/open-telemetry/opentelemetry-js/pull/6332) @vitorvasc
+
 ### :bug: Bug Fixes
 
 * fix(exporter-prometheus): add missing `@opentelemetry/semantic-conventions` dependency [#6330](https://github.com/open-telemetry/opentelemetry-js/pull/6330) @omizha